code clean based on findbugs and so on
git-svn-id: http://hazelcast.googlecode.com/svn/trunk@1399 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07

+++ hazelcast/src/main/java/com/hazelcast/jmx/ManagementService.java
@@ -183,7 +183,7 @@ public class ManagementService {
 
         public void run() {
             synchronized (this) {
-                average = events / interval;
+                average = (double) events / interval;
                 events = 0;
                 min = average < min ? average : min;
                 max = average > max ? average : max;
